fix(focal-point): load block library via schema-free shared loader
imageFocalPoint runs in both the edit and canvas editors, but it read the
block library through the edit-side da-library loader, whose eager HTML
parsing reads window.view.state.schema. In the canvas editor window.view
is undefined, so the focal-point icon failed to initialize.

Extract the canvas block-library loader (which stores raw DOM and needs no
schema) into blocks/shared/block-library.js and point imageFocalPoint at it.
Canvas call sites now import from shared; the canvas-only AEM Assets plugin
injection stays in canvas and is layered on in getCanvasToolPanelViews.
The shared module sources config helpers from blocks/shared/utils.js,
dropping the nx dependency.
