+++ blocks/canvas/ew-block-library-modal/ew-block-library-modal.js
@@ -1,10 +1,10 @@
 import { LitElement, html, nothing } from 'da-lit';
 import { getNx } from '../../../scripts/utils.js';
 import {
-  loadBlockLibrary,
   getItemPreviewUrl,
   getPreviewStatus,
 } from '../ew-panel-extensions/helpers.js';
+import { loadBlockLibrary } from '../../shared/block-library.js';
 
 const nx = getNx();
 await import(`${nx}/blocks/shared/dialog/dialog.js`);
