+++ blocks/canvas/ew-panel-extensions/ew-panel-library.js
@@ -1,7 +1,6 @@
 import { LitElement, html, nothing } from 'da-lit';
 import { getNx } from '../../../scripts/utils.js';
 import {
-  fetchBlocks,
   fetchItems,
   insertBlock,
   insertText,
@@ -9,6 +8,7 @@ import {
   getPreviewStatus,
   getItemPreviewUrl,
 } from './helpers.js';
+import { fetchBlocks } from '../../shared/block-library.js';
 import { getExtensionsBridge } from '../editor-utils/extensions-bridge.js';
 
 const { loadStyle, hashChange } = await import(`${getNx()}/utils/utils.js`);
