+++ blocks/edit/prose/plugins/imageFocalPoint.js
@@ -1,7 +1,8 @@
 import { Plugin, PluginKey } from 'da-y-wrapper';
 import inlinesvg from '../../../shared/inlinesvg.js';
 import { openFocalPointDialog } from './focalPointDialog.js';
-import { loadLibrary } from '../../da-library/helpers/helpers.js';
+import { loadBlockLibrary } from '../../../shared/block-library.js';
+import getPathDetails from '../../../shared/pathDetails.js';
 import { getTableInfo, isInTableCell } from './tableUtils.js';
 
 const imageFocalPointKey = new PluginKey('imageFocalPoint');
@@ -12,10 +13,9 @@ async function getBlocksData() {
   if (!blocksDataPromise) {
     blocksDataPromise = (async () => {
       try {
-        const libraryList = await loadLibrary();
-        const blocksInfo = libraryList.find((l) => l.name === 'blocks');
-        if (!blocksInfo?.loadItems) return [];
-        return await blocksInfo.loadItems;
+        const { org, site } = getPathDetails();
+        const { blocks } = await loadBlockLibrary(org, site);
+        return blocks;
       } catch (error) {
         // eslint-disable-next-line no-console
         console.warn('Failed to load blocks data for focal point:', error);
